% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beams.R
\docType{data}
\name{beams}
\alias{beams}
\title{The beams data set}
\value{
A data frame
}
\description{
Ten wooden beams had
their Strength measured together with their specific Gravity and Moisture.
}
\examples{
data(beams)
}
\keyword{datasets}

