% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{drphil}
\alias{drphil}
\title{Dr Phil's data set}
\usage{
data(drphil)
}
\value{
A data frame
}
\description{
Dr Phil comes to see you with his data. He believes that IQ can be
predicted by the number of years education. Dr Phil does not differentiate
between primary, secondary and tertiary education. He has four variables:
IQ, AgeBegin, AgeEnd, TotalYears
}
\examples{
data(drphil)
}
\keyword{datasets}

