\docType{data}
\name{graves}
\alias{graves}
\title{Graves data set}
\value{
  A data frame
}
\description{
  The data are from 101 consecutive patients attending a
  combined thyroid-eye clinic. The patients have an
  endocrine disorder, Graves' Ophthalmopathy, which affects
  various aspects of their eyesight. The ophthalmologist
  measures various aspects of their eyesight and constructs
  an overall index of how the disease affects their
  eyesight. This is the Ophthalmic Index (OI) given in the
  dataset. The age of the patient and their sex are also
  recorded. In practice, and as this is a chronic condition
  which can be ameliorated but not cured, the OI would be
  monitored at successive clinic visits to check on the
  patient's progress. However, these data are obtained at
  presentation. We are interested in how OI changes with
  age and gender.
}
\examples{
data(graves)
}
\keyword{datasets}

