% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ratfeed}
\alias{ratfeed}
\alias{ratfeed2}
\alias{ratfeed3}
\title{Ratfeed data set}
\usage{
data(ratfeed)
}
\value{
A data frame
}
\description{
The \code{ratfeed} data set. An example of the factorial ANOVA design.
The \code{ratfeed2} and \code{ratfeed3} datasets are similar to the original data,
but are used to illustrate variable encoding.
}
\examples{
data(ratfeed)
}
\keyword{datasets}

