\docType{package}
\name{nclRmodelling-package}
\alias{nclRmodelling}
\alias{nclRmodelling-package}
\title{R courses at Newcastle University}
\description{
  Functions and datasets used in for external R courses at
  Newcastle University
}
\author{
  \email{colin.gillespie@newcastle.ac.uk}
}
\keyword{package}

